GOPHORA-v2/backend/requirements.txt
# Core FastAPI
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv

# Authentication
passlib
argon2-cffi
python-jose[cryptography]

# Firebase
firebase-admin

# LangChain & AI Models
langchain
langchain-google-genai
langchain-openai
langchain-community
google-generativeai

# Web Scraping
requests
beautifulsoup4
scrapy
selenium
lxml
html2text

# NLP & ML
spacy
transformers
sentence-transformers

# Task Scheduling
apscheduler

# Utilities
python-multipart
aiohttp
